# backend/static/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "Lobster",
    "description": "Go to /login/ to obtain your token. Authorize with the following value: 'Token {your_token}'"
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "security": [
    {
      "ApiKeyAuth": [

      ]
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization"
      }
    }
  },
  "paths": {
    "/courses/": {
      "get": {
        "tags": [
          "Courses"
        ],
        "summary": "Gets a list of courses on lobster",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "Courses"
        ],
        "summary": "Adds a course to lobster",
        "requestBody": {
          "description": "Course Post Object",
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "semester": {
                    "type": "string"
                  },
                  "year": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "OK"
          }
        }
      }
    },
    "/courses/{name}/{semester}/{year}/": {
      "get": {
        "tags": [
          "Courses"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "schema": {
              "type": "string"
            },
            "required": true
          },
          {
            "in": "path",
            "name": "semester",
            "schema": {
              "type": "string"
            },
            "required": true
          },
          {
            "in": "path",
            "name": "year",
            "schema": {
              "type": "integer"
            },
            "required": true
          }
        ],
        "summary": "Gets details on a course on lobster",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/courses/{courseid}/": {
      "get": {
        "tags": [
          "Courses"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "courseid",
            "schema": {
              "type": "integer"
            },
            "required": true
          }
        ],
        "summary": "Gets details on a course on lobster",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "patch": {
        "tags": [
          "Courses"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "courseid",
            "schema": {
              "type": "integer"
            },
            "required": true
          }
        ],
        "requestBody": {
          "description": "Course Post Object",
          "required": false,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "semester": {
                    "type": "string"
                  },
                  "year": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        },
        "summary": "Edits a course on lobster",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Courses"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "courseid",
            "schema": {
              "type": "integer"
            },
            "required": true
          }
        ],
        "summary": "Deletes a course on lobster",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/courses/{courseid}/copy/": {
      "post": {
        "tags": [
          "Courses"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "courseid",
            "schema": {
              "type": "integer"
            },
            "required": true
          }
        ],
        "summary": "Copies an existing course on lobster. Copies exercises and course admins but does not copy other users",
        "responses": {
          "201": {
            "description": "OK"
          }
        }
      }
    },
    "/exercises/": {
      "post": {
        "tags": [
          "Exercises"
        ],
        "summary": "Adds a course to lobster",
        "requestBody": {
          "description": "Exercise Post Object",
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file ": {
                    "type": "object",
                    "properties": {
                      "filename": {
                        "type": "string"
                      },
                      "contents": {
                        "type": "string"
                      }
                    }
                  },
                  "courseid": {
                    "type": "integer"
                  },
                  "name": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "OK"
          }
        }
      }
    },
    "/courses/{courseid}/exercises/": {
      "get": {
        "tags": [
          "Exercises"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "courseid",
            "schema": {
              "type": "integer"
            },
            "required": true
          }
        ],
        "summary": "Gets a exercises associated with a course on lobster",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/exercises/{exerciseid}/": {
      "get": {
        "tags": [
          "Exercises"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "exerciseid",
            "schema": {
              "type": "integer"
            },
            "required": true
          }
        ],
        "summary": "Gets details on an exercise on lobster",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Exercises"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "exerciseid",
            "schema": {
              "type": "integer"
            },
            "required": true
          }
        ],
        "summary": "Gets details on an exercise on lobster",
        "responses": {
          "204": {
            "description": "OK"
          }
        }
      }
    },
    "/exercises/{exerciseid}/starter_files/{filename}/": {
      "get": {
        "tags": [
          "Exercises"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "exerciseid",
            "schema": {
              "type": "integer"
            },
            "required": true
          },
          {
            "in": "path",
            "name": "filename",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "summary": "Gets a exercises associated with a course on lobster",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/exercises/{exerciseid}/starter_files/": {
      "get": {
        "tags": [
          "Exercises"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "exerciseid",
            "schema": {
              "type": "integer"
            },
            "required": true
          }
        ],
        "summary": "Gets starter_files for an exercise on lobster",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "patch": {
        "tags": [
          "Exercises"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "exerciseid",
            "schema": {
              "type": "integer"
            },
            "required": true
          }
        ],
        "requestBody": {
          "description": "Course Post Object",
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "filename": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "summary": "Removes a starter_file for an exercise on lobster",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "Exercises"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "exerciseid",
            "schema": {
              "type": "integer"
            },
            "required": true
          }
        ],
        "summary": "Adds starter_files for an exercise on lobster",
        "requestBody": {
          "description": "Exercise Post Object",
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "filename": {
                    "type": "string"
                  },
                  "file_contents": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "OK"
          }
        }
      }
    },
...
  }
}